refactor(infra): split ObsidianMcpServerAdapter into per-tool-family registrars
Closes #243

ObsidianMcpServerAdapter.ts was 900 LOC and tripped the max-lines
warn floor (350). The five per-family register*Tools functions
already existed at module scope; move each into its own file under
src/infrastructure/obsidian/mcp/ so the adapter class owns only the
HTTP server and use-case wiring.

No public API change to ObsidianMcpServerPort consumers.

- src/infrastructure/obsidian/mcp/shared.ts
- src/infrastructure/obsidian/mcp/registerVaultAndFeatureTools.ts
- src/infrastructure/obsidian/mcp/registerWorkflowTools.ts
- src/infrastructure/obsidian/mcp/registerMetadataTools.ts
- src/infrastructure/obsidian/mcp/registerLinksTools.ts
- src/infrastructure/obsidian/mcp/registerCanvasTools.ts
- src/infrastructure/obsidian/mcp/registerBasesTools.ts
- src/infrastructure/obsidian/mcp/index.ts

Adapter file is now 129 LOC; every split file is under 170 LOC.
